0.0.3: rewrite to prevent holding onto processed data; some API changes
Functions + macro meant for public consumption in the current version:
  * parse
  * parse-string
  * process-json
  * define-json-processor

The lazy tree representation is gone from the public interface for
now; Jackson's parse events streams are now handled directly (the
parse and parse-string functions now produce seqs of parse events in
Clojure-friendly form). Note that the lazy tree infrastructure is
still being used to perform JSON-to-Clojure conversions for the
benefit of callback functions.